Make LineInfo track visit_counts instead of a simple bool.
This is a precursor to adding cache-grind support as an output format for the coverage grinder, allowing it to be used as a line profiler.

BUG=

Review URL: https://codereview.appspot.com/6816053

git-svn-id: http://sawbuck.googlecode.com/svn/trunk/syzygy@1204 15e8cca8-e42c-11de-a347-f34a4f72eb7d

+++ grinder/lcov_writer.h
@@ -1,4 +1,4 @@
-// Copyright 2012 Google Inc.
+// Copyright 2012 Google Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -102,7 +102,7 @@ class LcovWriter {
   struct CoverageInfo;  // Forward declaration.
 
   // A map of line numbers to execution counts.
-  typedef std::map<size_t, size_t> LineExecutionCountMap;
+  typedef std::map<size_t, uint32> LineExecutionCountMap;
   // A map of file names to coverage information.
   typedef std::map<std::string, CoverageInfo> SourceFileCoverageInfoMap;
